pyproject.toml
[tool.poetry]
name = "tusdatos"
version = "0.1.0"
description = ""
authors = ["Sebastian Rios Sabogal <sebaxtianrioss@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
fastapi = "0.111.0"
uvicorn = "0.29.0"
selenium = "4.20.0"
python-dotenv = "1.0.1"
loguru = "0.7.2"
icecream = "2.1.3"
pendulum = "3.0.0"
requests = "2.31.0"
fastapi-cache2 = {extras = ["memcache"], version = "0.2.1"}
pydantic-settings = "2.2.1"
sqlalchemy = "2.0.29"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
passlib = {extras = ["bcrypt=4.1.2"], version = "1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
